refactor: split prompts from pipeline
